// apps/web/src/lib/api/generated/types/createPlacementDtoFeeType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

/**
 * Fee type; defaults to PERCENTAGE
 */
export type CreatePlacementDtoFeeType = typeof CreatePlacementDtoFeeType[keyof typeof CreatePlacementDtoFeeType];


export const CreatePlacementDtoFeeType = {
  PERCENTAGE: 'PERCENTAGE',
  FLAT: 'FLAT',
} as const;