// apps/web/src/lib/api/generated/types/interviewEntityOutcome.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type InterviewEntityOutcome = typeof InterviewEntityOutcome[keyof typeof InterviewEntityOutcome];


export const InterviewEntityOutcome = {
  SCHEDULED: 'SCHEDULED',
  PENDING: 'PENDING',
  PASSED: 'PASSED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
} as const;