// apps/web/src/lib/api/generated/types/submissionEntityStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type SubmissionEntityStatus = typeof SubmissionEntityStatus[keyof typeof SubmissionEntityStatus];


export const SubmissionEntityStatus = {
  SUBMITTED: 'SUBMITTED',
  INTERVIEWING: 'INTERVIEWING',
  REJECTED: 'REJECTED',
  PLACED: 'PLACED',
} as const;